Pedigree likelihood

Calculates various forms of pedigree likelihoods based on paramlink functions.

likelihood.linkdat(
  x,
  locus1,
  locus2 = NULL,
  theta = NULL,
  startdata = NULL,
  eliminate = 0,
  logbase = NULL,
  loop_breakers = NULL,
  ...
)

likelihood.singleton(x, locus1, logbase = NULL, ...)

Arguments

x

a linkdat object, a singleton object, or a list of such objects. In likelihood_LINKAGE, x must be a linkdat object, with x$model different from NULL.

locus1

a marker object compatible with x. If x is a list, then locus1 must be a list of corresponding marker objects.

locus2

either NULL, the character 'disease', or a markerfb object compatible with x. See Details.

theta

the recombination rate between locus1 and locus2 (in likelihood_LINKAGE: between the marker and the disease locus). To make biological sense theta should be between 0 and 0.5.

startdata

for internal use. linkage computations with few-allelic markers.

eliminate

mostly for internal use: a non-negative integer indicating the number of iterations in the internal genotype-compatibility algorithm. Positive values can save time if partialmarker is non-empty and the number of alleles is large.

logbase

a numeric, or NULL. If numeric the log-likelihood is returned, with logbase as basis for the logarithm.

loop_breakers

a numeric containing IDs of individuals to be used as loop breakers. If NULL, automatic selection of loop breakers will be performed.

...

further arguments.

Value

The likelihood of the data. If the parameter logbase is a positive number, the output is log(likelihood, logbase).

Details

All likelihoods are calculated using the Elston-Stewart algorithm.