Linkdat objects

This function has been obtained from paramlink package, no longer maintained. Functions to create and display 'linkdat' objects.

Usage

linkdat(
  ped,
  model = NULL,
  map = NULL,
  dat = NULL,
  freq = NULL,
  annotations = NULL,
  missing = 0,
  header = FALSE,
  checkped = TRUE,
  verbose = TRUE,
  ...
)

singletonfb(id, sex = 1, famid = 1, verbose = FALSE, ...)

# S3 method for linkdat
print(x, ..., markers)

# S3 method for linkdat
summary(object, ...)

# S3 method for linkdat
subset(x, subset = x$orig.ids, ..., markers = seq_len(x$nMark))

Arguments

ped

a matrix, data.frame or a character with the path to a pedigree file in standard LINKAGE format. (See details)

model

either a linkdat.model object (typically y$model for some linkdat object y), or a single integer with the following meaning: 1 = autosomal dominant; 2 = autosomal recessive; 3 = X-linked dominant; 4 = X-linked recessive. In each of these cases, the disease is assumed fully penetrant and the disease allele frequency is set to 0.00001. If model=NULL, no model is set.

map

a character with the path to a map file in MERLIN format, or NULL. If non-NULL, a dat file must also be given (next item).

dat

a character with the path to a dat file in MERLIN format, or NULL. (Only needed if map is non-NULL.)

freq

a character with the path to a allele frequency file in MERLIN (short) format, or NULL. If NULL, all markers are interpreted as equifrequent.

annotations

a list (of the same length and in the same order as the marker columns in x) of marker annotations. If this is non-NULL, then all of map, dat, freq should be NULL.

missing

the character (of length 1) used for missing alleles. Defaults to '0'.

header

a logical, relevant only if ped points to a ped file: If TRUE, the first line of the ped file is skipped.

checkped

a logical. If FALSE, no checks for pedigree errors are performed.

verbose

a logical: verbose output or not.

...

further arguments.

id, sex

single numerics describing the individual ID and gender of the singletonfb.

famid

a numeric: the family ID of the singletonfb.

x, object

a linkdat object.

markers

a numeric indicating which markers should be included/printed.

subset

a numeric containing the individuals in the sub-pedigree to be extracted. NB: No pedigree checking is done here, so make sure the subset form a meaningful, closed pedigree.

Value

A linkdat object, or a list of linkdat objects. A linkdat object is essentially a list with the following entries, some of which can be NULL.

pedigree

data.frame with 5 columns (ID, FID, MID, SEX, AFF) describing the pedigree in linkage format. (NB: Internal labeling used.)

orig.ids

the original individual id labels.

nInd

the number of individuals in the pedigree.

founders

vector of the founder individuals. (NB: Internal labeling used.)

nonfounders

vector of the nonfounder individuals (NB: Internal labeling used.)

hasLoops

a logical: TRUE if the pedigree is inbred.

subnucs

list containing all (maximal) nuclear families in the pedigree. Each nuclear family is given as a vector of the form c(pivot, father, mother, child1, ...), where the pivot is either the id of the individual linking the nuclear family to the rest of the pedigree, or 0 if there are none. (NB: Internal labeling used.)

markerdata

a list of marker objects.

nMark

the number of markers.

available

a numeric vector containing IDs of available individuals. Used for simulations and plots.

model

a linkdat.model object, essentially a list containing the model parameters. See setModel for details.

loop_breakers

a matrix with original loop breaker ID's in the first column and their duplicates in the second column. This is set by breakLoopsfb.

Details

The file (or matrix or data.frame) ped must describe one or several pedigrees in standard LINKAGE format, i.e. with the following columns in correct order:

1 Family id (optional) (FAMID)

2 Individual id (ID),

3 Father id (FID),

4 Mother id (MID),

5 Gender (SEX): 1 = male, 2 = female,

6 Affection status (AFF): 1 = unaffected, 2 = affected, 0 = unknown,

7 First allele of first marker,

8 Second allele of first marker,

9 First allele of second marker,

a.s.o.

Only columns 2-6 are mandatory. The first column is automatically interpreted as family id if it has repeated elements.

Internally the individuals are relabelfbed as 1,2,..., but this should rarely be of concern to the end user. Some pedigree checking is done, but it is recommended to plot the pedigree before doing any analysis.

Details on the formats of map, dat and frequency files can be found in the online MERLIN tutorial: http://csg.sph.umich.edu/abecasis/Merlin/

A singletonfb is a special linkdat object whose pedigree contains 1 individual. The class attribute of a singletonfb is c('singletonfb', 'linkdat')

See also