Marker functions

Functions for setting and manipulating marker genotypes for 'linkdat' objects compatible with fbnet. It was extracted from paramlink package, no longer maintained.

Usage

markerfb(
  x,
  ...,
  allelematrix,
  alleles = NULL,
  afreq = NULL,
  missing = 0,
  chrom = NA,
  pos = NA,
  name = NA,
  mutmat = NULL
)

addMarkerfb(x, m, ...)

SetMarkersfb(x, m, annotations = NULL, missing = 0)

modifyMarkerfb(x, marker, ids, genotype, alleles, afreq, chrom, name, pos)

getMarkersfb(
  x,
  markernames = NULL,
  chroms = NULL,
  fromPos = NULL,
  toPos = NULL
)

removeMarkersfb(
  x,
  markers = NULL,
  markernames = NULL,
  chroms = NULL,
  fromPos = NULL,
  toPos = NULL
)

swapGenotypesfb(x, ids)

modifyMarkerfbMatrix(x, ids, new.alleles)

Arguments

x

a linkdat object

...

an even number of vectors, indicating individuals and their genotypes. See examples.

allelematrix

a matrix with one row per pedigree member and two columns per marker, containing the alleles for a single marker.

alleles

a numeric or character vector containing allele names.

afreq

a numerical vector with allele frequencies. An error is given if they don't sum to 1 (rounded to 3 decimals).

missing

a numeric - or character - of length 1, indicating the code for missing alleles.

chrom

NA or an integer (the chromosome number of the marker).

pos

NA or a non-negative real number indicating the genetic position (in cM) of the marker.

name

NA or a character (the name of the marker).

mutmat

a mutation matrix, or a list of two such matrices named 'female' and 'male'. The matrix/matrices must be square, with the allele labels as dimnames, and each row must sum to 1 (after rounding to 3 decimals).

m

a marker object or a matrix with alleles. (In SetMarkersfb this matrix can contain data of several markers.)

annotations

a list of marker annotations.

marker, markers

a numeric indicating which marker(s) to use/modify.

ids

a numeric indicating individual(s) to be modified. In swapGenotypesfb this must have length 2.

genotype

a vector of length 1 or 2, containing the genotype to be given the ids individuals. See examples.

markernames

NULL or a character vector.

chroms

NULL or a numeric vector of chromosome numbers.

fromPos, toPos

NULL or a single numeric.

new.alleles

a numerical matrix of dimensions length(ids), 2*x$nMark. Entries refer to the internal allele numbering.

Value

The marker function returns an object of class marker: This is a numerical 2-column matrix with one row per individual, and attributes 'alleles' (a character vector with allele names), 'nalleles' (the number of alleles) and 'missing' (the input symbol for missing marker alleles), 'chrom' (chromosome number), 'name' (marker identifier), 'pos' (chromosome position in cM).

For addMarker, SetMarkersfb, removeMarkersfb, modifyMarkerfb, modifyMarkerfbMatrix and swapGenotypesfb, a linkdat object is returned, whose markerdata element has been set/modified.

For getMarkersfb a numeric vector containing marker numbers (i.e. their indices in x$markerdata) for the markers whose 'name' attribute is contained in markernames, 'chrom' attribute is contained in chroms, and 'pos' attribute is between from and to. NULL arguments are skipped, so getMarkersfb(x) will return seq_len(x$nMark) (i.e. all markers).

See also