Modify the pedigree of 'linkdat' objects

Functions to modify the pedigree of a 'linkdat' object.

addOffspring(
  x,
  father,
  mother,
  noffs,
  ids = NULL,
  sex = 1,
  aff = 1,
  verbose = TRUE
)

removeIndividualsfb(x, ids, verbose = TRUE)

trim(x, keep = c("available", "affected"), return.ids = FALSE, verbose = TRUE)

relabelfb(x, new, old)

Arguments

x

A linkdat object

father, mother

Integers indicating the IDs of parentsfb. If missing, a new founder individual is created (whose ID will be 1+the largest ID already in the pedigree).

noffs

A single integer indicating the number of offspringfb to be created.

ids

individuals

sex, aff

Integer vectors indicating the gender and affection statuses of the offspringfb to be created (recycled if less than noffs elements).

verbose

A logical: Verbose output or not.

keep

A character, either 'available' (trimming the pedigree for unavailable members) or 'affected' (trimming for unaffected members).

return.ids

A logical. If FALSE, the trimmed pedigree is returned as a new linkdat object. If TRUE, a vector containing the IDs of 'removable' individuals is returned

new

a numeric containing new labels to replace those in old.

old

a numeric containing ID labels to be replaced by those in new. If missing, old is set to x$orig.ids, i.e. all members in their original order.

Value

The modified linkdat object.

Details

When removing an individual, all descendantsfb are also removed as well as founders remaining without offspringfb.

See also