Pedigree subsets

Utility functions for 'linkdat' objects, mainly for extracting various pedigree information.

offspringfb(x, id, original.id = TRUE)

spousesfb(x, id, original.id = TRUE)

related.pairs(
  x,
  relation = c("parentsfb", "siblingsfb", "grandparentsfbfb", "nephews_niecesfb",
    "cousins", "spousesfb", "unrelatedfb"),
  available = F,
  interfam = c("none", "founders", "all"),
  ...
)

unrelatedfb(x, id, original.id = TRUE)

leavesfb(x)

parentsfb(x, id, original.id = TRUE)

grandparentsfbfb(x, id, degree = 2, original.id = TRUE)

siblingsfb(x, id, half = NA, original.id = TRUE)

cousins(x, id, degree = 1, removal = 0, half = NA, original.id = TRUE)

nephews_niecesfb(x, id, removal = 1, half = NA, original.id = TRUE)

ancestorsfb(x, id)

descendantsfb(x, id, original.id = TRUE)

Arguments

x

a linkdat object. In related.pairs possibly a list of linkdat objects.

id

a numerical ID label.

original.id

a logical indicating whether 'id' refers to the original ID label or the internal labeling.

relation

one of the words (possibly truncated) parentsfb, siblingsfb, grandparentsfbfb, nephews_niecesfb, cousins, spousesfb, unrelatedfb.

available

a logical, if TRUE only pairs of available individuals are returned.

interfam

one of the words (possibly truncated) none, founders or all, specifying which interfamiliar pairs should be included as unrelatedfb in the case where x is a list of several pedigrees. If none, only intrafamiliar pairs are considered; if founders all interfamiliar pairs of (available) founders are included; if all, all interfamiliar (available) pairs are included.

...

further parameters

degree

a non-negative integer.

half

a logical or NA. If TRUE (resp FALSE), only half (resp. full) siblingsfb/cousins/nephews/nieces are returned. If NA, both categories are included.

removal

a non-negative integer

Value

For ancestorsfb(x,id), a vector containing the ID's of all ancestorsfb of the individual id. For descendantsfb(x,id), a vector containing the ID's of all descendantsfb (i.e. children, grandchildren, a.s.o.) of individual id.

The functions cousins, grandparentsfbfb, nephews_niecesfb, offspringfb, parentsfb, siblingsfb, spousesfb, unrelatedfb, each returns an integer vector containing the ID's of all pedigree members having the specified relationship with id.

For related.pairs a matrix with two columns. Each row gives of a pair of pedigree members with the specified relation. If the input is a list of multiple pedigrees, the matrix entries are characters of the form 'X-Y' where X is the family ID and Y the individual ID of the person.

For leavesfb, a vector of IDs containing all pedigree members without children.