Pedigree loops

Functions for identifying, breaking and restoring loops in pedigrees.

pedigreeLoops(x)

breakLoopsfb(x, loop_breakers = NULL, verbose = TRUE)

tieLoopsfb(x)

findLoopBreakersfb(x)

findLoopBreakersfb2(x)

Arguments

x

a linkdat object.

loop_breakers

either NULL (resulting in automatic selection of loop breakers) or a numeric containing IDs of individuals to be used as loop breakers.

verbose

a logical: Verbose output or not?

Value

For breakLoopsfb, a linkdat object in which the indicated loop breakers are duplicated. The returned object will also have a non-null loop_breakers entry, namely a matrix with the IDs of the original loop breakers in the first column and the duplicates in the second.

For tieLoopsfb, a linkdat object in which any duplicated individuals (as given in the x$loop_breakers entry) are merged. For any linkdat object x, the call tieLoopsfb(breakLoopsfb(x)) should return x.

For pedigreeLoops, a list containing all inbreeding loops (not marriage loops) found in the pedigree. Each loop is represented as a list with elements 'top', a 'bottom' individual, 'pathA' (individuals forming a path from top to bottom) and 'pathB' (creating a different path from top to bottom, with no individuals in common with pathA). Note that the number of loops reported here counts all closed paths in the pedigree and will in general be larger than the genus of the underlying graph.

For findLoopBreakersfb and findLoopBreakersfb2, a numeric vector of individual ID's.

Details

Most of paramlink's handling of pedigree loops is done under the hood - using the functions described here - without need for explicit action from end users. When a linkdat object x is created, an internal routine detects if the pedigree contains loops, in which case x$hasLoops is set to TRUE. In analyses of x where loops must be broken (e.g. lod score computation or marker simulation), this is done automatically by calling breakLoopsfb.

In some cases with complex inbreeding, it can be instructive to plot the pedigree after breaking the loops. Duplicated individuals are plotted with appropriate labels (see examples).

The function findLoopBreakersfb identifies a set of individuals breaking all inbreeding loops, but not marriage loops. These require more machinery for efficient detection, and paramlink does this is a separate function, findLoopBreakersfb2, utilizing methods from the igraph package. Since this is rarely needed for most users, igraph is not imported when loading paramlink, only when findLoopBreakersfb2 is called.

In practice, breakLoopsfb first calls findLoopBreakersfb and breaks at the returned individuals. If the resulting linkdat object still has loops, findLoopBreakersfb2 is called to break any marriage loops.